codegeex/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "a CodeGeeX proxy"

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "^1.51.3"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
